refactor: data types
