+++ Blokus/DataTypes/Cell.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum Cell: Codable {
+  case empty
+  case occupied(pieceID: String, owner: PlayerColor)
+}
